requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2

# Deep Learning
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
pytorch-lightning==2.1.2

# Real-time Data & Trading
websockets==12.0
ccxt==4.1.55
python-binance==1.0.19

# Database & Caching
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
loguru==0.7.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
aiofiles==23.2.0
asyncio==3.4.3
httpx==0.25.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Visualization
plotly==5.17.0
dash==2.14.2
